Extracted ChoreDetails to its own component
